Rename originDetails to originResponse
The proper name for this key is originResponse, although it should have
optional id, message, and code.

We currently have a concept of a BraintreeOriginResponse but it requires
a code and message without an id.

We'll have to sort this out but for now renaming the key solves the
immediate issue.